Detect the cycle in a graph using BFS
This program takes an adjacency matrix representation of a graph as input and uses BFS to detect if a cycle exists starting from a specified start node. It returns true if a cycle is detected and false otherwise. The algorithm works by using a visited array to keep track of nodes that have been visited during BFS. If a node is already visited and it is not the parent of the current node being visited, then a cycle is detected. The program uses a queue to perform BFS traversal and returns the result as a boolean value.

Overall, this program and algorithm are a useful way to detect cycles in a graph using BFS. By keeping track of visited nodes and checking for cycles during traversal, the program is able to detect cycles in a variety of graphs.