Breadth First Search Using Colors
Breadth First Search Using Colors
In this code, we first define a 2D array adjMatrix to represent the adjacency matrix of the graph. We then call the bfs method with the matrix and the source vertex as input.

Inside the bfs method, we first initialize the arrays color, prev, and d with default values, and create an empty queue q.

We then mark the source vertex as GREY, set its distance to 0, and add it to the queue. We then enter a loop that continues until the queue is empty. In each iteration of the loop, we dequeue a vertex u from the queue, and iterate over its neighbors v. If a neighbor v is white (unvisited), we mark it as GREY, set its distance and parent, and add it to the queue. After all the neighbors of u have been processed, we mark u as BLACK.

Finally, we output the values of color, prev, and d.

Note that in this code, we are using the string "WHITE", "GREY", and "BLACK" to represent the colors of vertices. We are also using the value Integer.MAX_VALUE to represent an infinite distance.